requirements.txt
autopep8==1.5.5
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
gspread==3.7.0
httplib2==0.19.0
idna==2.10
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
